Update Sidenav.scss with improved padding and background color
The Sidenav.scss file has been updated to improve the padding and background color of the navigation sidebar. The padding-top value has been adjusted to 138px for better alignment. The background color of the nav-logo element has been changed to #222746. These changes enhance the visual appearance and usability of the navigation sidebar.