Update Sidenav and Home.scss for improved styling and usability
Adjust the width and height of the Sidenav component to improve its appearance and usability. Change the height to 100vh and the color to match the nightowl theme. Add a sticky position and top value of 0 to keep the Sidenav fixed at the top of the page. Update the Home.scss file to include lazy loading for images and adjust the styling of the appointment section. These changes enhance the visual appeal and user experience of the navigation sidebar and home page.